CSP を unsafe-inline から nonce ベースに移行する
- middleware.ts を新規作成（Next.js App Router 対応）
  - リクエストごとに UUID から Base64 ノンスを生成
  - script-src に 'nonce-{nonce}' を設定し unsafe-inline を廃止
  - style-src は 'unsafe-inline' を維持（Leaflet のインライン CSS 注入に必要）
  - Content-Security-Policy を Report-Only から強制ヘッダーに昇格
  - Supabase Auth セッション更新も統合
- next.config.js から CSP 設定を削除（middleware に移管）
